+++ packages/aesthetic-adapter-aphrodite/tests/UnifiedAdapter.test.js
@@ -3,7 +3,7 @@
 import { StyleSheet, StyleSheetTestUtils } from 'aphrodite';
 import UnifiedAphroditeAdapter from '../src/UnifiedAdapter';
 import {
-  FONT_ROBOTO,
+  FONT_ROBOTO_FLAT_SRC,
   KEYFRAME_FADE,
   SYNTAX_UNIFIED_FULL,
   SYNTAX_AT_RULES,
@@ -27,7 +27,7 @@ describe('aesthetic-adapter-aphrodite/UnifiedAdapter', () => {
 
   it('transforms style declarations into class names', () => {
     expect(instance.transform('component', SYNTAX_UNIFIED_FULL)).toEqual({
-      button: 'button_808fk2',
+      button: 'button_1fihu6i',
     });
   });
 
@@ -40,7 +40,7 @@ describe('aesthetic-adapter-aphrodite/UnifiedAdapter', () => {
         borderRadius: 4,
         display: 'inline-block',
         cursor: 'pointer',
-        fontFamily: FONT_ROBOTO,
+        fontFamily: [FONT_ROBOTO_FLAT_SRC],
         fontWeight: 'normal',
         lineHeight: 'normal',
         whiteSpace: 'nowrap',
@@ -49,7 +49,7 @@ describe('aesthetic-adapter-aphrodite/UnifiedAdapter', () => {
         backgroundColor: '#337ab7',
         verticalAlign: 'middle',
         color: 'rgba(0, 0, 0, 0)',
-        animationName: KEYFRAME_FADE,
+        animationName: [KEYFRAME_FADE],
         animationDuration: '.3s',
         ':hover': {
           backgroundColor: '#286090',
@@ -81,7 +81,7 @@ describe('aesthetic-adapter-aphrodite/UnifiedAdapter', () => {
   it('supports font faces', () => {
     expect(instance.convert(SYNTAX_FONT_FACE)).toEqual({
       font: {
-        fontFamily: FONT_ROBOTO,
+        fontFamily: [FONT_ROBOTO_FLAT_SRC],
         fontSize: 20,
       },
     });
@@ -90,7 +90,7 @@ describe('aesthetic-adapter-aphrodite/UnifiedAdapter', () => {
   it('supports animations', () => {
     expect(instance.convert(SYNTAX_KEYFRAMES)).toEqual({
       animation: {
-        animationName: KEYFRAME_FADE,
+        animationName: [KEYFRAME_FADE],
         animationDuration: '3s, 1200ms',
         animationIterationCount: 'infinite',
       },