+++ packages/aesthetic-adapter-fela/src/UnifiedAdapter.js
@@ -5,13 +5,22 @@
  */
 
 import UnifiedSyntax from 'aesthetic/unified';
-import { injectAtRules, injectFallbacks, injectRuleByLookup } from 'aesthetic-utils';
+import {
+  injectFallbacks,
+  injectFontFaces,
+  injectKeyframes,
+  injectMediaQueries,
+} from 'aesthetic-utils';
 import FelaAdapter from './NativeAdapter';
 
 import type { Renderer } from 'fela'; // eslint-disable-line
-import type { StyleDeclarationMap, TransformedStylesMap, CSSStyle } from '../../types';
-
-const SRC_PATTERN = /url\((?:'|")?([^()'"]+)(?:'|")?\)/ig;
+import type {
+  FontFace,
+  Keyframe,
+  StyleDeclaration,
+  StyleDeclarations,
+  TransformedDeclarations,
+} from '../../types';
 
 export default class UnifiedFelaAdapter extends FelaAdapter {
   syntax: UnifiedSyntax;
@@ -19,58 +28,61 @@ export default class UnifiedFelaAdapter extends FelaAdapter {
   constructor(fela: Renderer, options?: Object = {}) {
     super(fela, options);
 
-    this.syntax = new UnifiedSyntax();
-    this.syntax
-      .on('declaration', this.onDeclaration)
-      .on('fontFace', this.onFontFace)
-      .on('keyframe', this.onKeyframe);
+    this.syntax = new UnifiedSyntax()
+      .on('declaration', this.handleDeclaration)
+      .on('fontFace', this.handleFontFace)
+      .on('keyframe', this.handleKeyframe);
   }
 
-  convert(declarations: StyleDeclarationMap): StyleDeclarationMap {
+  convert(declarations: StyleDeclarations): StyleDeclarations {
     return this.syntax.convert(declarations);
   }
 
-  transform(styleName: string, declarations: StyleDeclarationMap): TransformedStylesMap {
+  transform<T: Object>(styleName: string, declarations: T): TransformedDeclarations {
     return super.transform(styleName, this.convert(declarations));
   }
 
-  extractFontSources(source: string): string[] {
-    return (source.match(SRC_PATTERN) || []).map((url: string) => (
-      url.replace(/"|'|url\(|\)/g, '')
-    ));
-  }
-
-  onDeclaration = (setName: string, properties: CSSStyle) => {
+  handleDeclaration = (selector: string, properties: StyleDeclaration) => {
     // Font faces
+    // http://fela.js.org/docs/basics/Fonts.html
+    // http://fela.js.org/docs/basics/Renderer.html#renderfont
     if ('fontFamily' in properties) {
-      injectRuleByLookup(properties, 'fontFamily', this.syntax.fontFaceNames, true);
+      injectFontFaces(properties, this.syntax.fontFacesCache, {
+        join: true,
+      });
     }
 
     // Animation keyframes
+    // http://fela.js.org/docs/basics/Keyframes.html
+    // http://fela.js.org/docs/basics/Renderer.html#renderkeyframe
     if ('animationName' in properties) {
-      injectRuleByLookup(properties, 'animationName', this.syntax.keyframeNames, true);
+      injectKeyframes(properties, this.syntax.keyframesCache, {
+        join: true,
+      });
     }
 
     // Media queries
-    if (this.syntax.mediaQueries[setName]) {
-      injectAtRules(properties, '@media', this.syntax.mediaQueries[setName]);
+    // http://fela.js.org/docs/basics/Rules.html#3-media-queries
+    if (this.syntax.mediaQueries[selector]) {
+      injectMediaQueries(properties, this.syntax.mediaQueries[selector]);
     }
 
     // Fallbacks
-    if (this.syntax.fallbacks[setName]) {
-      injectFallbacks(properties, this.syntax.fallbacks[setName]);
+    // https://github.com/rofrischmann/fela/tree/master/packages/fela-plugin-fallback-value#example
+    if (this.syntax.fallbacks[selector]) {
+      injectFallbacks(properties, this.syntax.fallbacks[selector]);
     }
   };
 
-  onFontFace = (setName: string, familyName: string, properties: CSSStyle) => {
-    this.syntax.fontFaceNames[familyName] = this.fela.renderFont(
-      familyName,
-      this.extractFontSources(String(properties.src)),
-      properties,
-    );
+  handleFontFace = (selector: string, familyName: string, fontFaces: FontFace[]) => {
+    this.syntax.fontFacesCache[familyName] = fontFaces.map((face) => {
+      const { src, ...props } = face;
+
+      return this.fela.renderFont(familyName, src, props);
+    });
   }
 
-  onKeyframe = (setName: string, animationName: string, properties: CSSStyle) => {
-    this.syntax.keyframeNames[animationName] = this.fela.renderKeyframe(() => properties);
+  handleKeyframe = (selector: string, animationName: string, keyframe: Keyframe) => {
+    this.syntax.keyframesCache[animationName] = this.fela.renderKeyframe(() => keyframe);
   };
 }