+++ packages/aesthetic-utils/src/index.js
@@ -1,21 +1,26 @@
 /**
  * @copyright   2017, Miles Johnson
  * @license     https://opensource.org/licenses/MIT
+ * @flow
  */
 
 import createStyleElement from './createStyleElement';
-import injectAtRules from './injectAtRules';
+import formatFontFace from './formatFontFace';
 import injectFallbacks from './injectFallbacks';
-import injectRuleByLookup from './injectRuleByLookup';
+import injectFontFaces from './injectFontFaces';
+import injectKeyframes from './injectKeyframes';
+import injectMediaQueries from './injectMediaQueries';
 import isObject from './isObject';
 import isPrimitive from './isPrimitive';
 import toArray from './toArray';
 
 export {
   createStyleElement,
-  injectAtRules,
+  formatFontFace,
   injectFallbacks,
-  injectRuleByLookup,
+  injectFontFaces,
+  injectKeyframes,
+  injectMediaQueries,
   isObject,
   isPrimitive,
   toArray,