+++ packages/aesthetic/tests/UnifiedSyntax.test.js
@@ -39,7 +39,7 @@ describe('aesthetic/UnifiedSyntax', () => {
 
       expect(instance.convert({
         '@font-face': {
-          mrroboto: FONT_ROBOTO,
+          Roboto: FONT_ROBOTO,
         },
         '@keyframes': {
           fade: KEYFRAME_FADE,
@@ -47,7 +47,7 @@ describe('aesthetic/UnifiedSyntax', () => {
       })).toEqual({});
 
       expect(instance.fontFaces).toEqual({
-        Roboto: FONT_ROBOTO,
+        Roboto: [FONT_ROBOTO],
       });
       expect(instance.keyframes).toEqual({
         fade: KEYFRAME_FADE,
@@ -171,7 +171,7 @@ describe('aesthetic/UnifiedSyntax', () => {
 
       expect(() => (
         instance.extractFontFaces('foo', {
-          mrroboto: FONT_ROBOTO,
+          Roboto: FONT_ROBOTO,
         }, GLOBAL)
       )).toThrowError('Font face "Roboto" has already been defined.');
     });
@@ -180,11 +180,11 @@ describe('aesthetic/UnifiedSyntax', () => {
       expect(instance.fontFaces).toEqual({});
 
       instance.extractFontFaces('foo', {
-        mrroboto: FONT_ROBOTO,
+        Roboto: FONT_ROBOTO,
       }, GLOBAL);
 
       expect(instance.fontFaces).toEqual({
-        Roboto: FONT_ROBOTO,
+        Roboto: [FONT_ROBOTO],
       });
     });
   });