+++ packages/aesthetic-adapter-css-modules/CHANGELOG.md
@@ -1,3 +1,7 @@
+# 2.0.0
+#### 🛠 Internal
+* Rewritten Flowtype definitions.
+
 # 1.6.2 - 11/10/17
 #### 🛠 Internal
 * Tested against React 16.1.